+++ patches/src/main/kotlin/app/morphe/patches/reddit/misc/fix/signature/SpoofSignaturePatch.kt
@@ -10,7 +10,7 @@ private const val EXTENSION_CLASS_DESCRIPTOR =
 @Suppress("unused")
 val spoofSignaturePatch = bytecodePatch(
     name = "Spoof signature",
-    description =  "Spoofs the signature of the app to fix notification issues."
+    description = "Spoofs the signature of the app to fix notification issues."
 ) {
     compatibleWith(COMPATIBILITY_REDDIT)