+++ patches/src/main/kotlin/app/morphe/patches/youtube/layout/toolbar/Fingerprints.kt
@@ -4,8 +4,7 @@ import app.morphe.patcher.Fingerprint
 import com.android.tools.smali.dexlib2.AccessFlags
 
 internal object ToolBarPatchFingerprint : Fingerprint(
-    accessFlags = listOf(AccessFlags.PRIVATE, AccessFlags.STATIC),
-    custom = { method, _ ->
-        method.name == "hookToolBar"
-    }
+    definingClass = EXTENSION_CLASS_DESCRIPTOR,
+    name = "hookToolBar",
+    accessFlags = listOf(AccessFlags.PRIVATE, AccessFlags.STATIC)
 )